Basic kotlin integration tests for querying coturn servers
The tests just check the authenticated and non-authenticated flows
as a smoke test.

I tried to use 'quarkus.hibernate-orm.sql-load-script', but this only
worked when I disabled flyway for the tests. I couldn't find an
analogous option for flyway, so inserting the test data in the test
itself seemed the most straightforward option.